- nmake warning
git-svn-id: svn://rdo.rk9.bmstu.ru/var/svn/root@12309 23f926e6-80f1-49ca-85a0-60d4a27bb3ac

+++ rdo/packages/converter/smr2rdox/trunk/rdoparser_lexer.cpp
@@ -48,7 +48,7 @@ int RDOLexer::LexerInput(PTR(char) buf, int max_size)
 	}
 	else
 	{
-		return m_yyin->gcount();
+		return static_cast<int>(m_yyin->gcount());
 	}
 #endif
 }
